More documentation, general improvements and domain whitelist.
For the first time, I tested spider.js against a real website, not just
apache2 or nginx directory listings. This helped me to spot some
suboptimal areas and make improvements.

The visitedUris and pendingUris checks really help to reduce the amount
of duplicate/recursive work required when crawling for links.

The testing highlighted the importance of having a domain whitelist
feature, otherwise the spider will (as expected) go and recursively
trawl every site it finds a.hrefs to.

spider.js actually uncovered a non-obvious 404 error on my personal
website.

By design, the script doesn't currently download HTML, CSS, JavaScript
and other related assets, so it's not currently suitable for full
website mirroring and archiving, but the "code logic" base seems to be
coming along. Maybe one day!

* Added/improved misc. code docs.
* Added warning about deprecated libs/utils/frameworks and documented
  low vs. high risk use cases.
* spider() function now checks the uri against the visitedUris
  collection at the start of the call and skips if it matches.
* spider() function, at the start of the call, now checks the uri
  against the domainWhitelist collection and skips if NOT matched.
* Page found a.hrefs.host is now checked against the domainWhitelist
  collection and skipped if NOT matched.
* Page found a.hrefs.href are checked against the pendingUris collection
  and skipped if already pending.
* Page found a.hrefs.href is checked against the visitedUris collection
  and skipped if already pending.

The last 3 points should provide a reasonable improvement in processing
performance, as the new logic reduces the amount of recursion and work
performed.

* Captured knowledge about the CasperJS package, including the fact that
  it uses its own JavaScript environment. This includes difficulties I
  encountered during implementation and a link dump for anyone who wants
  to follow in my footsteps.
* Added additional documentation and a backstory/history section.
* Added domainWhitelist feature.
* Renamed uriFilter variable to uriBlacklist.
* domainWhitelist and uriBlacklist both use case-insensitive RegEx,
  previously uriFilter was case-sensitive.
* Added basic DEBUG boolean logic. Enabling this increases the verbosity
  of the console logs and enables debug mode on CasperJS/PhantomJS.
* Added isInArray() utility function.
* Added some metrics for tracking what happens during a crawl.
* Changed the logic/prototype of the spider() function to no longer take
  an argument. Now the parsed uri object is moved from the pendingUris
  collection at the start of each spider() call. This logic has been moved
  out of the checkPending() pending function.